[Bronze III] Title: 알고리즘 수업 - 알고리즘의 수행 시간 3, Time: 100 ms, Memory: 14224 KB -BaekjoonHub

+++ 백준/Bronze/24264. 알고리즘 수업 － 알고리즘의 수행 시간 3/알고리즘 수업 － 알고리즘의 수행 시간 3.java
@@ -5,7 +5,7 @@ public class Main{
     public static void main(String[] argss) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         
-        long n = Integer.parseInt(br.readLine());
+        long n = Long.parseLong(br.readLine());
 
         System.out.println(n*n);
         System.out.println(2);
